# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  model.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  model.Token:
    properties:
      access_token:
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  model.UploadToken:
    properties:
      upload_token:
        type: string
    type: object
  model.User:
    properties:
      account:
        example: act_001
        type: string
      account_number:
        example: 1001
        type: integer
      anonymous_probe:
        example: false
        type: boolean
      email:
        example: user@example.com
        type: string
      is_active:
        example: true
        type: boolean
      is_admin:
        example: false
        type: boolean
      nickname:
        example: 小明
        type: string
      qq_number:
        example: 12345678
        type: integer
      upload_token:
        example: token_xyz
        type: string
      user_id:
        type: integer
      username:
        example: user123
        type: string
      uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - email
    - username
    type: object
...
  request.ChangePasswordRequest:
    properties:
      new_password:
        example: newpass456
        minLength: 6
        type: string
      old_password:
        example: oldpass123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  request.CreateUserRequest:
    properties:
      account:
        example: act_001
        type: string
      account_number:
        example: 1001
        type: integer
      anonymous_probe:
        example: false
        type: boolean
      email:
        example: user@example.com
        type: string
      is_active:
        example: true
        type: boolean
      is_admin:
        example: false
        type: boolean
      nickname:
        example: 小明
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
      qq_number:
        example: 12345678
        type: integer
      upload_token:
        example: token_xyz
        type: string
      username:
        example: user123
        type: string
      uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - email
    - password
    - username
    type: object
  request.ResetPasswordRequest:
    properties:
      new_password:
        example: newpass456
        minLength: 6
        type: string
      username:
        example: targetuser
        type: string
    required:
    - new_password
    - username
    type: object
...
  request.UpdateUserRequest:
    properties:
      account:
        type: string
      account_number:
        type: integer
      anonymous_probe:
        type: boolean
      nickname:
        type: string
      qq_number:
        type: integer
      uuid:
        type: string
    type: object
host: api.prp.icel.site
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 'This is the API documentation for the Paradigm: Reboot Prober service.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'Paradigm: Reboot Prober API'
  version: "2"
paths:
...
  /user/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user and return JWT token
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Token'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: Login user
      tags:
      - user
  /user/me:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update the profile of the currently authenticated user
      parameters:
      - description: User update info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Update current user info
      tags:
      - user
  /user/me/password:
    put:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password change info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - user
  /user/me/upload-token:
    post:
      description: Generate a new upload token for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UploadToken'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Refresh upload token
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: Register a new user
      tags:
      - user
  /user/reset-password:
    post:
      consumes:
      - application/json
      description: Reset a user's password by admin
      parameters:
      - description: Password reset info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: Reset user password (Admin only)
      tags:
      - user
schemes:
- https
swagger: "2.0"